third_party/blink/web_tests/external/wpt/dom/nodes/Document-createElementNS-expected.txt

This is a testharness.js-based test.
Found 12 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] createElementNS test in HTML document: undefined,"f::oo","NAMESPACE_ERR"
  assert_throws_dom: function "function() {doc.createElementNS(namespace, qualifiedName) }" threw object "InvalidCharacterError: Failed to execute 'createElementNS' on 'Document': The qualified name provided ('f::oo') has an empty local name." that is not a DOMException NAMESPACE_ERR: property "code" is equal to 5, expected 14
[FAIL] createElementNS test in XML document: undefined,"f::oo","NAMESPACE_ERR"
  assert_throws_dom: function "function() {doc.createElementNS(namespace, qualifiedName) }" threw object "InvalidCharacterError: Failed to execute 'createElementNS' on 'Document': The qualified name provided ('f::oo') has an empty local name." that is not a DOMException NAMESPACE_ERR: property "code" is equal to 5, expected 14
[FAIL] createElementNS test in XHTML document: undefined,"f::oo","NAMESPACE_ERR"
  assert_throws_dom: function "function() {doc.createElementNS(namespace, qualifiedName) }" threw object "InvalidCharacterError: Failed to execute 'createElementNS' on 'Document': The qualified name provided ('f::oo') has an empty local name." that is not a DOMException NAMESPACE_ERR: property "code" is equal to 5, expected 14
[FAIL] createElementNS test in HTML document: "http://example.com/","f:o:o",null
  assert_equals: element.localName expected "o:o" but got "o"
[FAIL] createElementNS test in XML document: "http://example.com/","f:o:o",null
  assert_equals: element.localName expected "o:o" but got "o"
[FAIL] createElementNS test in XHTML document: "http://example.com/","f:o:o",null
  assert_equals: element.localName expected "o:o" but got "o"
[FAIL] createElementNS test in HTML document: "http://example.com/","f::oo",null
  Failed to execute 'createElementNS' on 'Document': The qualified name provided ('f::oo') has an empty local name.
[FAIL] createElementNS test in XML document: "http://example.com/","f::oo",null
  Failed to execute 'createElementNS' on 'Document': The qualified name provided ('f::oo') has an empty local name.
[FAIL] createElementNS test in XHTML document: "http://example.com/","f::oo",null
  Failed to execute 'createElementNS' on 'Document': The qualified name provided ('f::oo') has an empty local name.
[FAIL] createElementNS test in HTML document: "http://example.com/","prefix::local",null
  Failed to execute 'createElementNS' on 'Document': The qualified name provided ('prefix::local') has an empty local name.
[FAIL] createElementNS test in XML document: "http://example.com/","prefix::local",null
  Failed to execute 'createElementNS' on 'Document': The qualified name provided ('prefix::local') has an empty local name.
[FAIL] createElementNS test in XHTML document: "http://example.com/","prefix::local",null
  Failed to execute 'createElementNS' on 'Document': The qualified name provided ('prefix::local') has an empty local name.
Harness: the test ran to completion.
